+++ src/kiezellisp-lib/reader-handlers.cs
@@ -322,11 +322,6 @@ namespace Kiezel
             }
         }
 
-        public static object ReadInfixHandler(LispReader stream, string ch, int arg)
-        {
-            return stream.ParseInfixExpression();
-        }
-
         public static object ReadLineCommentHandler(LispReader stream, char ch)
         {
             // ;